Move from _widget to _main_widget
