+++ src/cavendish_particle_tracks/_main_widget.py
